Refactor retired-actor cleanup into spec-driven hosted service
Extract the one-off ChannelRuntime cleanup into a generic, spec-driven
mechanism that runs on every pod startup and is intrinsically idempotent.

- IRetiredActorSpec abstraction in Aevatar.Foundation.Abstractions.Maintenance:
  modules declare retired type tokens, well-known targets, optional dynamic
  discovery, and optional read-model cleanup.
- Generic RetiredActorCleanupHostedService in Aevatar.Foundation.Runtime.Hosting
  iterates every registered IRetiredActorSpec under a per-spec lease
  (__maintenance:retired-actor-cleanup:{specId}). No "completed forever"
  marker — each spec runs every startup and is a no-op when no targets
  match.
- ChannelRuntime/Device/Scheduled each ship their own IRetiredActorSpec
  alongside their DI extension; the user-agent catalog walker lives with
  Scheduled. Adding a new retired type only requires registering a new
  spec — Mainnet.Host.Api stays untouched.
- Mainnet.Host.Api wires AddRetiredActorCleanup() in place of the old
  module-specific hosted service.
